+++ src/app/core/services/local-storage.service.ts
@@ -1,0 +1,23 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class LocalStorageService {
+
+  getItem(key:string): string | null{
+    return localStorage.getItem(key);
+  }
+
+  setItem(key: string, value: string){
+    localStorage.setItem(key, value);
+  }
+
+  removeItem(key: string){
+    localStorage.removeItem(key);
+  }
+
+  clear(): void {
+    localStorage.clear();
+  }
+}