+++ src/app/product/components/product-list/product-list.component.ts
@@ -17,6 +17,6 @@ export class ProductListComponent {
   }
 
   onAddToCart(item: Product){
-    this.cartService.addToCart(item);
+    this.cartService.addProduct(item);
   }
 }