+++ src/app/shared/shared.module.ts
@@ -1,18 +1,21 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HighlightDirective } from './directives/highlight.directive';
+import { ClickStyleDirective } from './directives/click-style.directive';
 
 
 
 @NgModule({
   declarations: [
-    HighlightDirective
+    HighlightDirective,
+    ClickStyleDirective
   ],
   imports: [
     CommonModule
   ],
   exports: [
-    HighlightDirective
+    HighlightDirective,
+    ClickStyleDirective
   ]
 })
 export class SharedModule { }